+++ lib/stackup/utils.rb
@@ -1,0 +1,21 @@
+module Stackup
+
+  # Generates diffs of data.
+  #
+  module Utils
+
+    def normalize_data(data)
+      case data
+      when Hash
+        pairs = data.sort.map { |k,v| [k, normalize_data(v)] }
+        Hash[pairs]
+      when Array
+        pairs = data.map { |x| normalize_data(x) }
+      else
+        data
+      end
+    end
+
+  end
+
+end